monoBundle support + code Refactoring
