+++ src/definitions/FrameworkSettings.ts
@@ -1,3 +1,4 @@
+import { Module } from "./Module";
 import { StatusManagerOptions } from "./StatusManagerOptions";
 
 export interface FrameworkSettings {
@@ -11,4 +12,5 @@ export interface FrameworkSettings {
     };
     defaultPrefix?: string;
     statusOptions?: StatusManagerOptions;
+    srcMode?: 'multiBundle' | 'monoBundle' | undefined;
 }
