+++ src/index.ts
@@ -28,9 +28,18 @@ import { ZumitoFramework } from './ZumitoFramework.js';
 import * as discord from 'discord.js';
 import { EventParameters } from './definitions/parameters/EventParameters.js';
 import { ServiceContainer } from './services/ServiceContainer.js';
+import { GuildDataGetter } from './services/GuildDataGetter.js';
+import { MemberPermissionChecker } from './services/MemberPermissionChecker.js';
+import { CommandParser } from './services/CommandParser.js';
+import { SlashCommandRefresher } from './services/SlashCommandRefresher.js';
 
 ServiceContainer.addService(TextFormatter, []);
 ServiceContainer.addService(EmojiFallback, []);
+ServiceContainer.addService(GuildDataGetter, [ZumitoFramework.name]);
+ServiceContainer.addService(MemberPermissionChecker, []);
+ServiceContainer.addService(CommandParser, []);
+ServiceContainer.addService(SlashCommandRefresher, [ZumitoFramework.name]);
+
 
 export {
     ZumitoFramework,
@@ -58,5 +67,8 @@ export {
     StatusManagerOptions,
     discord,
     EventParameters,
-    ServiceContainer
+    ServiceContainer,
+    GuildDataGetter,
+    SlashCommandRefresher,
+    CommandParser,
 };
