+++ src/services/RecursiveObjectMerger.ts
@@ -1,0 +1,20 @@
+export class RecursiveObjectMerger {
+    public static merge(obj1, obj2) {
+        for (const p in obj2) {
+            try {
+                // Property in destination object set; update its value.
+                if (obj2[p].constructor == Object) {
+                    obj1[p] = this.merge(obj1[p], obj2[p]);
+                } else {
+                    obj1[p] = obj2[p];
+                }
+            } catch (e) {
+                // Property in destination object not set; create it and set its value.
+                obj1[p] = obj2[p];
+            }
+        }
+    
+        return obj1;
+    }
+    
+}
